+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggplot2 (development version)
 
+* New `strip.spacing` theme element, determining the distance between panels
+  and strips (@teunbrand, #5935)
 * (internal) The plot's layout now has a coord parameter that is used to 
   prevent setting up identical panel parameters (#5427)
 * (internal) rearranged the code of `Facet$draw_panels()` method (@teunbrand).